pyproject.toml
[project]
name = "sbom_grader"
version = "0.1.0"
description = "SBOM Grader"
authors = [
    {name = "mszymutk", email = "mszymutk@redhat.com"},
]
dependencies = [
    "jsonschema>=4.23.0",
    "pyyaml>=6.0.2",
    "spdx-tools>=0.8.3",
    "packageurl-python>=0.16.0",
]
requires-python = ">=3.12"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"


[tool.pdm]
distribution = true

[tool.pdm.dev-dependencies]
dev = [
    "black>=24.10.0",
    "tox>=4.22.0",
]
